refactor: Profile 페이지의 left container 컴포넌트 분리
